HealthGoal and CustomHealthGoal modify service, add HealthGoal testing

+++ SEP490_FTCDHMM_API.Application/Dtos/NutrientDtos/NutrientTarget/NutrientTargetRequest.cs
@@ -8,7 +8,7 @@
         public decimal? MaxValue { get; set; }
         public decimal? MinEnergyPct { get; set; }
         public decimal? MaxEnergyPct { get; set; }
-        public decimal Weight { get; set; } = 1m;
+        public int Weight { get; set; } = 1;
     }
 
 }
